# api/v1alpha1/provider/openapi.yaml
openapi: 3.0.4
info:
  contact: {}
  description: |
    DCM Service Provider API - Registration and management of Service Providers.
    This API implements the Service Provider registration flow as defined in the
    DCM enhancement proposal.
  title: Service Provider API
  version: v1alpha1
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html

servers:
  - url: /api/v1alpha1

tags:
  - name: provider
    description: Service Provider management operations
  - name: health
    description: Health check operations

paths:
  /health:
    get:
      tags:
        - health
      summary: Health check
      operationId: getHealth
      description: Health check for DCM Service Provider API
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Health'

  /providers:
    get:
      tags:
        - provider
      summary: List all providers
      operationId: listProviders
      description: Returns a list of registered service providers with optional filtering
      parameters:
        - name: type
          in: query
          description: Filter providers by service type
          schema:
            type: string
        - name: max_page_size
          in: query
          description: Maximum number of results per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 100
        - name: page_token
          in: query
          description: Token for pagination
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderList'
        '400':
          description: Bad request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: Unexpected error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags:
        - provider
      summary: Register a Service Provider
      operationId: createProvider
      description: |
        Register a new service provider or update an existing one.
        Registration is idempotent:
        - If name does not exist, a new SP entry is created
        - If name exists with same/no providerID, the entry is updated
        - If name exists with different providerID, registration fails (conflict)
        - If providerID exists with different name, registration fails (conflict)
      parameters:
        - name: id
          in: query
          description: Optional provider ID for idempotent registration
          schema:
            type: string
            pattern: '^[a-z0-9]([a-z0-9-]{0,61}[a-z0-9])?$'
            minLength: 1
            maxLength: 63
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Provider'
      responses:
        '200':
          description: Provider updated (re-registration)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Provider'
        '201':
          description: Provider created (new registration)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Provider'
        '400':
          description: Invalid input
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Conflict - name or providerID already in use
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Validation exception
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: Unexpected error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Error'

  /providers/{providerId}:
    get:
      tags:
        - provider
      summary: Get a provider
      operationId: getProvider
      description: Get a service provider by its unique ID
      parameters:
        - $ref: '#/components/parameters/ProviderIdPath'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Provider'
        '400':
          description: Invalid ID supplied
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Provider not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: Unexpected error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags:
        - provider
      summary: Update a Service Provider
      operationId: applyProvider
      description: |
        Update an existing service provider.
        The providerID in the path must match an existing provider.
      parameters:
        - $ref: '#/components/parameters/ProviderIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Provider'
      responses:
        '200':
          description: Provider updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Provider'
        '400':
          description: Invalid input
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Provider not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Conflict - name already in use by another provider
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: Unexpected error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - provider
      summary: Delete a service Provider
      operationId: deleteProvider
      description: Remove a service provider from the registry
      parameters:
        - $ref: '#/components/parameters/ProviderIdPath'
      responses:
        '204':
          description: Provider deleted successfully
        '400':
          description: Invalid ID supplied
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Provider not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: Unexpected error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  parameters:
    ProviderIdPath:
      name: providerId
      in: path
      required: true
      schema:
        type: string
        pattern: '^[a-z0-9]([a-z0-9-]{0,61}[a-z0-9])?$'
        minLength: 1
        maxLength: 63
      description: Unique identifier of the provider
      example: "123e4567-e89b-12d3-a456-426614174000"
  schemas:
    ProviderMetadata:
      type: object
      description: Additional metadata about the provider
      additionalProperties: true
      properties:
        region_code:
          type: string
          description: Geographic region code where the provider operates
          example: "us-east-1"
        zone:
          type: string
          description: Availability zone or datacenter identifier
          example: "datacenter-b"
        status:
          type: string
          description: Current operational status of the provider
          example: "healthy"
        resources:
          $ref: '#/components/schemas/ResourceCapacity'

    ResourceCapacity:
      type: object
      description: Resource capacity information
      properties:
        total_cpu:
          type: integer
          description: Total CPU cores available
          example: 200
        total_memory:
          type: string
          description: Total memory available
          example: "1TB"
        total_storage:
          type: string
          description: Total storage available
          example: "2TB"
        total_node:
          type: integer
          description: Total number of nodes
          example: 100

    Provider:
      type: object
      description: Full provider resource representation
      x-aep-resource:
        type: serviceprovider.dcm.io/provider
        singular: provider
        plural: providers
        patterns:
          - providers/{provider_id}
      required:
        - name
        - endpoint
        - service_type
        - schema_version
      properties:
        id:
          type: string
          pattern: '^[a-z0-9]([a-z0-9-]{0,61}[a-z0-9])?$'
          minLength: 1
          maxLength: 63
          readOnly: true
          description: Unique identifier for the Service Provider
          example: "123e4567-e89b-12d3-a456-426614174000"
        path:
          type: string
          readOnly: true
          description: Resource path identifier
          example: "providers/123e4567-e89b-12d3-a456-426614174000"
        name:
          type: string
          description: Unique name of the Service Provider
          example: "kubevirt-123"
        display_name:
          type: string
          description: Human-readable display name for the provider
          example: "KubeVirt Service Provider"
        endpoint:
          type: string
          format: uri
          description: Full endpoint URL where the provider API is accessible
          example: "https://sp1.example.com/api/v1alpha1/vm"
        service_type:
          type: string
          description: Type of service this provider offers
          example: "vm"
        schema_version:
          type: string
          description: Schema version of the service type the SP supports
          pattern: "^v[0-9]+(alpha|beta)?[0-9]*$"
          example: "v1alpha1"
        operations:
          type: array
          items:
            type: string
          description: List of operations supported for this service type
          example: ["create", "delete", "update"]
        metadata:
          $ref: '#/components/schemas/ProviderMetadata'
        health_status:
          type: string
          description: Health status of the provider
          example: "ready"
          readOnly: true
        create_time:
          type: string
          format: date-time
          readOnly: true
          description: Timestamp when the provider was first registered
        update_time:
          type: string
          format: date-time
          readOnly: true
          description: Timestamp when the provider was last updated

    ProviderList:
      type: object
      description: Paginated list of providers
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/Provider'
        next_page_token:
          type: string
          description: Token for retrieving the next page of results
          example: "eyJpZCI6IjEyM2U0NTY3LWU4OWItMTJkMy1hNDU2LTQyNjYxNDE3NDAwMCJ9"

    Error:
      type: object
      description: RFC 7807 compliant error response
      required:
        - type
        - title
      properties:
        type:
          type: string
          format: uri-reference
          description: URI reference identifying the error type
          example: "https://dcm.example.com/errors/invalid-input"
        title:
          type: string
          description: Short human-readable summary of the problem
          example: "Invalid Input"
        status:
          type: integer
          description: HTTP status code
          example: 400
        detail:
          type: string
          description: Human-readable explanation specific to this occurrence
          example: "The 'name' field is required but was not provided"
        instance:
          type: string
          format: uri-reference
          description: URI reference for this specific error occurrence
          example: "/errors/123e4567-e89b-12d3-a456-426614174000"

    Health:
      type: object
      description: Health status singleton resource
      x-aep-resource:
        singular: health
        plural: healths
        singleton: true
        type: serviceprovider.dcm.io/health
        patterns:
          - health
      properties:
        status:
          type: string
          description: Health status
          example: "healthy"
        path:
          type: string
          description: Canonical path of the resource
          example: "health"
          readOnly: true